HomeTest/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
openai==1.14.0
httpx==0.25.2
python-dotenv==1.0.0
